OneDrive/Desktop/adk/requirements.txt
yfinance
psutil
litellm
google-generativeai
python-dotenv